comment some debug messages and fix bug
When we were collecting graphs before the end of the level, it is
possible that there are less graphs left in cur_level than slaves. In
that case, after collecting the graphs, we wouldn't be able to send a
graph to every slave, meaning that when we collected the graphs at the
end of a level, the slaves we didn't send graphs to wouldn't send a
slave request, causing the master to hang. Fix this by including a
check for this when we collect graphs mid-level